requirements.txt
numpy==1.17.2
pandas==0.25.1
scipy==1.3.1
apache-airflow==1.10.5